[IMP] sale_order_partial_invoice: Reorder and split files
