+++ sale_order_partial_invoice/wizard/sale_view.xml
@@ -2,30 +2,6 @@
 <openerp>
   <data>
 
-    <record id="view_order_line_tree" model="ir.ui.view">
-      <field name="name">sale.order.line.tree</field>
-      <field name="model">sale.order.line</field>
-      <field name="inherit_id" ref="sale.view_order_line_tree"/>
-      <field name="arch" type="xml">
-        <field name="product_uom_qty" position="after">
-          <field name="qty_invoiced" string="Inv. Qty"/>
-          <field name="qty_delivered" string="Shipped Qty"/>
-        </field>
-      </field>
-    </record>
-
-    <record id="view_order_line_form2" model="ir.ui.view">
-      <field name="name">sale.order.line.form2</field>
-      <field name="model">sale.order.line</field>
-      <field name="inherit_id" ref="sale.view_order_line_form2"/>
-      <field name="arch" type="xml">
-        <xpath expr="//field[@name='product_uom_qty']/.." position="after">
-          <field name="qty_invoiced" />
-          <field name="qty_delivered"/>
-        </xpath>
-      </field>
-    </record>
-
     <record id="view_sale_order_invoice_partially_form" model="ir.ui.view">
       <field name="name">sale.order.line.invoice.partially.form</field>
       <field name="model">sale.order.line.invoice.partially</field>
